Add macro to share adapter implementation for both validator and executor
