+++ nomos-services/data-availability/sampling/src/network/adapters/mod.rs
@@ -1,2 +1,6 @@
 #[cfg(feature = "libp2p")]
-pub mod libp2p;
+pub mod common;
+#[cfg(feature = "libp2p")]
+pub mod executor;
+#[cfg(feature = "libp2p")]
+pub mod validator;
